tests/providers/test_openai_client.py
"""
Unit tests for OpenAI client.

This module contains tests for the OpenAI client functionality.
"""

import unittest
import pytest
from src.providers import openai_client
from src.providers.exceptions import JsonParsingError
from tests.providers.fixtures.openai_responses import *


class TestOpenAIClient(unittest.TestCase):
    """Tests for the OpenAIClient class."""
    
    def setUp(self):
        """Set up test fixtures."""
        pass
    
    def tearDown(self):
        """Tear down test fixtures."""
        pass
    
    def test_basic_functionality(self):
        """Test basic functionality."""
        # TODO: Implement test
        pass


class TestJSONParsing(unittest.TestCase):
    """Tests for the JSONParsing class."""
    
    def setUp(self):
        """Set up test fixtures."""
        pass
    
    def tearDown(self):
        """Tear down test fixtures."""
        pass
    
    def test_basic_functionality(self):
        """Test basic functionality."""
        # TODO: Implement test
        pass


class TestErrorHandling(unittest.TestCase):
    """Tests for the ErrorHandling class."""
    
    def setUp(self):
        """Set up test fixtures."""
        pass
    
    def tearDown(self):
        """Tear down test fixtures."""
        pass
    
    def test_basic_functionality(self):
        """Test basic functionality."""
        # TODO: Implement test
        pass


class TestRetryLogic(unittest.TestCase):
    """Tests for the RetryLogic class."""
    
    def setUp(self):
        """Set up test fixtures."""
        pass
    
    def tearDown(self):
        """Tear down test fixtures."""
        pass
    
    def test_basic_functionality(self):
        """Test basic functionality."""
        # TODO: Implement test
        pass

if __name__ == '__main__':
    unittest.main()
