requirements.txt
google-generativeai
requests
python-dotenv
aiohttp
openai>=1.0.0
pyyaml
